fix(content-system): move app style-option table migration to 6.7
