+++ src/Core/Migration/V6_7/Migration1782423127AddAppContentSystemStyleOptionTable.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Core\Migration\V6_8;
+namespace Shopware\Core\Migration\V6_7;
 
 use Doctrine\DBAL\Connection;
 use Shopware\Core\Framework\Log\Package;
