+++ tests/migration/Core/V6_7/Migration1782423127AddAppContentSystemStyleOptionTableTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Migration\Core\V6_8;
+namespace Shopware\Tests\Migration\Core\V6_7;
 
 use Doctrine\DBAL\Connection;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -8,7 +8,7 @@ use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelLifecycleManager;
 use Shopware\Core\Framework\Util\Database\TableHelper;
-use Shopware\Core\Migration\V6_8\Migration1782423127AddAppContentSystemStyleOptionTable;
+use Shopware\Core\Migration\V6_7\Migration1782423127AddAppContentSystemStyleOptionTable;
 
 /**
  * @internal
